a lot of changes to sort entries by tags
including new models to act like dictionaries, not including actual
sorting. for now it just stores the data. This is likely to screw a lot
of things up and I don't really have time to work on it more tonight,
so feel free to revert if you want to work on something without dealing
with my new code. although unfortunately it will also change the
databases to it'll take more than just git pull and resetting the
servers to revert