typo in variable name, note on south migrations in extended description
because south migration operations happen directly through ssh any file
changes made by south do not show up on github. the only effect of this
as far as i can tell is that there are folders called migrations under
content, guardian, easy_thumbnails, userena, userena/contrib/umessages,
profiles and taggit that are will not be kept up to date on your local
copy. this shouldn't matter, but if it does that's why it's happening